Move code around to get rid of dependency cycle
